requirements.txt
pretty-midi==0.2.9
PyYAML==5.3.1
pypianoroll==1.0.4
tqdm==4.46.0
tensorboardX==2.1
SoundFile==0.10.3.post1
torch==1.6.0
librosa==0.8.1
joblib==0.14.1
audioread==2.1.9
